Moved pages to own module.
